Update HexFormat.cs
Properly handle end of file hex format line (:00000001FF). In the original code if this line is encountered the field type is not set, leaving it at the default (0 = Data), however Data field is null. This throws an exception down the line when this line's Data is accessed by the caller.
The checksum calculation appears to be non-spec, replaced with the proper one. The old one does not fail, but may not detect the mismatch.